Update Job_Zone_Reference.sql
I think this table is in the third normal form because it has a primary key that the other columns/rows depend on (it also mentions in the text itself the primary key for this table is job_zone). Also, in every row in the columns, there is no duplication of data.